requirements.txt
wxPython==4.0.7
PyYAML==6.0
charset-normalizer==2.0.12
